requirements/common.txt
Django==3.2.*
djangorestframework==3.13.*
djangorestframework-simplejwt==5.1.*
